sessao9/blog_project/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
pytest==8.0.0
pytest-django==4.7.0
freezegun==1.4.0
pytest-cov==4.1.0
